Added the current read out of the HV supplies
I had to change some subvis that contained the clusters we make for the
slow control. Since we also want to read out the current in the slow
loop. I wired the NIST time server subvi correctly (with the 4 years
substracted). And I made some icons for some subvis that didn't have one
yet.