Slow timer bugs fixed, Continuous NIST time syncing enabled and safeguards on arduino read buffer enabled
Fixed slow timer bug.  This bug involved syncing to the previous slow
time rather than the previous slow NIST time, causing time to not be
incremented.

Continuous NIST time syncing enabled for those who can read from their
NIST time server an unlimited number of times.

The data good? boolean was enabled and is used to ensure the string
buffer from the arduino is complete.  If it isn't, data is not recorded
or displayed.