Warning added if not connected to NIST timeserver
Upon initialization, ModuDAQ will attempt to contact the NIST timeserver
to get a timestamp to correlate fast and slow data.  If this fails, the
time will be generated locally, and a warning dialog will pop up.
Though time is generated locally, the warning just states that you
failed to contact the NIST timeserver, so your timestamps may not be
accurate (and this is not recommended for a science run).