Better error handling when getting NIST time and see arbitrary choice of waveform
Now the sGet NIST Time.vi does not return an error if the only NTP-related error was 56 (timeout, happens occasionally in normal use). Otherwise the error is returned. sTimer.vi passes out any errors from sGet NIST Time.vi and these can be caught where it is used in ModuHost.vi, which is now done in the slow control loop so that unexpected NTP errors can be displayed in the error out tab of ModuHost.vi

Also added a new control to the waveform watching tab so that it is now possible to see not just the first waveform in a cluster

Signed-off-by: Adam Brown <adambrown3456@gmail.com>